fix bug

+++ resources/views/posts/show.blade.php
@@ -62,13 +62,14 @@
                             <a href="{{ route('posts.top_up',$post->id) }}" class="btn btn-outline-success btn-sm" onclick="return confirm('確定要置頂？')"><i class="fas fa-sort-amount-up"></i> 置頂</a>
                         @endif
                     @endif
+
+                    @if(auth()->user()->id == $post->user_id or auth()->user()->admin ==1)
+                        <a href="{{ route('posts.edit',$post->id) }}" class="btn btn-outline-primary btn-sm"><i class="fas fa-edit"></i> 修改</a>
+                        <a href="#" class="btn btn-danger btn-sm" onclick="if(confirm('確定刪除？')) document.getElementById('delete').submit();else return false;"><i class="fas fa-trash"></i> 刪除</a>
+                        {{ Form::open(['route' => ['posts.destroy',$post->id], 'method' => 'DELETE','id'=>'delete','onsubmit'=>'return false;']) }}
+                        {{ Form::close() }}
+                    @endif
                 @endauth
-                @can('update',$post)
-                <a href="{{ route('posts.edit',$post->id) }}" class="btn btn-outline-primary btn-sm"><i class="fas fa-edit"></i> 修改</a>
-                <a href="#" class="btn btn-danger btn-sm" onclick="if(confirm('確定刪除？')) document.getElementById('delete').submit();else return false;"><i class="fas fa-trash"></i> 刪除</a>
-                {{ Form::open(['route' => ['posts.destroy',$post->id], 'method' => 'DELETE','id'=>'delete','onsubmit'=>'return false;']) }}
-                {{ Form::close() }}
-                @endcan
             </p>
 
             <hr>
